Split REST tests in 3 parts
