+++ tests/tas-restapi/pom.xml
@@ -29,9 +29,27 @@
 
     <profiles>
         <profile>
-            <id>run-tas-restapi-tests-with-ais</id>
+            <id>run-restapi-ais</id>
             <properties>
-                <suiteXmlFile>${project.basedir}/src/test/resources/restapi-ais-suite.xml</suiteXmlFile>
+                <suiteXmlFile>${project.basedir}/src/test/resources/test-suites/restapi-ais-suite.xml</suiteXmlFile>
+            </properties>
+        </profile>
+        <profile>
+            <id>run-restapi-part1</id>
+            <properties>
+                <suiteXmlFile>${project.basedir}/src/test/resources/test-suites/part1-suite.xml</suiteXmlFile>
+            </properties>
+        </profile>
+        <profile>
+            <id>run-restapi-part2</id>
+            <properties>
+                <suiteXmlFile>${project.basedir}/src/test/resources/test-suites/part2-suite.xml</suiteXmlFile>
+            </properties>
+        </profile>
+        <profile>
+            <id>run-restapi-part3</id>
+            <properties>
+                <suiteXmlFile>${project.basedir}/src/test/resources/test-suites/part3-suite.xml</suiteXmlFile>
             </properties>
         </profile>
     </profiles>
