+++ tests/tas-restapi/src/test/java/org/alfresco/rest/misc/FunctionalCasesTests.java
@@ -1,8 +1,9 @@
-package org.alfresco.rest;
+package org.alfresco.rest.misc;
 
 import org.alfresco.dataprep.CMISUtil;
 import org.alfresco.dataprep.CMISUtil.DocumentType;
 import org.alfresco.dataprep.SiteService.Visibility;
+import org.alfresco.rest.RestTest;
 import org.alfresco.rest.model.RestActivityModelsCollection;
 import org.alfresco.rest.model.RestCommentModel;
 import org.alfresco.rest.model.RestCommentModelsCollection;
