+++ tests/tas-restapi/src/test/resources/restapi-suite.xml
@@ -7,9 +7,16 @@
         <listener class-name="org.alfresco.utility.testng.OSTestMethodSelector"/>
 	</listeners>
 
-	<test name="Rest API Tests">
-		<packages>
-            <package name="org.alfresco.rest.*"/>
-		</packages>
-	</test>
+    <!--
+    The suites in "test-suites" folder are balanced according to execution time to make the CI/CD efficient.
+    Any new test should be allocated inside the existing package or a new package should be created and
+    added to the test suit. Currently the packages are created according to API groupings plus "misc" folder
+    with tests testing several APIs.
+    -->
+
+    <suite-files>
+        <suite-file path="test-suites/part1-suite.xml"/>
+        <suite-file path="test-suites/part2-suite.xml"/>
+        <suite-file path="test-suites/part3-suite.xml"/>
+    </suite-files>
 </suite>
