+++ tests/tas-restapi/src/test/resources/test-suites/part1-suite.xml
@@ -1,0 +1,20 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE suite SYSTEM "http://testng.org/testng-1.0.dtd">
+<suite name="REST API tests part1" preserve-order="true">
+	<listeners>
+        <listener class-name="org.alfresco.utility.report.HtmlReportListener"/>
+        <listener class-name="org.alfresco.utility.testrail.TestRailExecutorListener"/>
+        <listener class-name="org.alfresco.utility.testng.OSTestMethodSelector"/>
+	</listeners>
+
+    <test name="Part1">
+        <packages>
+            <package name="org.alfresco.rest.actions.*"/>
+            <package name="org.alfresco.rest.audit.*"/>
+            <package name="org.alfresco.rest.auth.*"/>
+            <package name="org.alfresco.rest.comments.*"/>
+            <package name="org.alfresco.rest.downloads.*"/>
+            <package name="org.alfresco.rest.favorites.*"/>
+        </packages>
+    </test>
+</suite>
